fix login bug and rename index to controller
