+++ src/main/ipc.js
@@ -14,7 +14,7 @@ const windows = require('./windows')
 const log = require('../logger').getLogger('main/ipc', true)
 const DeltaChat = (() => {
   try {
-    return require('./deltachat/index.js')
+    return require('./deltachat/controller')
   } catch (error) {
     log.critical('Fatal: The DeltaChat Module couldn\'t be loaded. Please check if all dependencies for deltachat-core are installed!', error)
     const { dialog } = require('electron')
